feat: 重构测试接口实现

+++ src/test/Test.service.ts
@@ -2,15 +2,24 @@ import * as fs from 'fs';
 import * as fse from 'fs-extra';
 import * as path from 'path';
 import * as shell from 'shelljs';
-import * as uuidv5 from 'uuid/v5';
+import * as uuidv4 from 'uuid/v4';
 import { Injectable } from '@nestjs/common';
 
 const SITESPEED = /sitespeedio\/browsertime/gi;
 const CONTEXT = fs.realpathSync(process.cwd());
 
+const getOutput = id => `./target/test/${id}`;
+const getOutputPath = id => path.resolve(CONTEXT, getOutput(id));
+const getInfoPath = id => path.resolve(CONTEXT, getOutput(id), 'info.json');
+const getBrowsertimePath = id =>
+  path.resolve(CONTEXT, getOutput(id), 'browsertime.json');
+const getHARPath = id =>
+  path.resolve(CONTEXT, getOutput(id), 'browsertime.har');
+
 @Injectable()
 export default class TestService {
   private queue = [];
+  private running: string = null;
 
   getTestingCount(): Promise<number> {
     return new Promise((resolve, reject) => {
@@ -36,26 +45,21 @@ export default class TestService {
     return this.queue.length;
   }
 
-  flushQueue(): void {
-    if (this.queue.length > 0) {
-      const current = this.queue[0];
-      if (!current.running) {
-        current.running = true;
-        current.run();
-      }
-    }
+  getPendingCountById(id): number {
+    return this.queue.indexOf(id);
   }
 
-  runTest(
-    url,
-  ): Promise<{ id: string; timestamp: number; browsertime: Object }> {
-    return new Promise(async (resolve, reject) => {
-      const run = () => {
-        const id = uuidv5(url, uuidv5.URL);
-        const timestamp = Date.now();
-        // TODO 优化目录结构
-        const output = `./target/test/${id}/${timestamp}`;
-        fse.ensureDirSync(path.resolve(CONTEXT, output));
+  async flushQueue() {
+    if (this.queue.length > 0 && !this.running) {
+      const id = this.queue[0];
+      this.running = id;
+      let info = null;
+      try {
+        info = await fse.readJson(getInfoPath(id));
+      } catch (error) {}
+      if (info && info.url) {
+        const output = getOutput(id);
+        const { url } = info;
         shell.exec(
           `docker run -e TZ=Asia/Shanghai --shm-size=1g --network=3g --rm -v "$(pwd)":/browsertime sitespeedio/browsertime ${url} --resultDir ${output}`,
           {
@@ -63,30 +67,85 @@ export default class TestService {
             silent: true,
           },
           (code, stdout, stderr) => {
-            const index = this.queue.indexOf(task);
-            this.queue.splice(index, 1);
-            this.flushQueue();
+            this.running = null;
             if (code === 0) {
-              const browsertime = fse.readJsonSync(
-                path.resolve(output, 'browsertime.json'),
-              );
-              resolve({
-                id,
-                timestamp,
-                browsertime,
-              });
+              this.queue.splice(0, 1);
+              this.flushQueue();
             } else {
-              reject(new Error(stderr));
+              // TODO 存储错误信息
+              setTimeout(() => {
+                this.flushQueue();
+              }, 3000);
             }
           },
         );
-      };
-      const task = {
-        running: false,
-        run,
-      };
-      this.queue.push(task);
-      this.flushQueue();
+      } else {
+        this.running = null;
+        this.queue.splice(0, 1);
+      }
+    }
+  }
+
+  async analysis(url): Promise<string> {
+    const id = uuidv4();
+    // TODO 优化输出目录（通过环境变量配置到项目外）
+    await fse.ensureDir(getOutputPath(id));
+    await fse.writeJson(getInfoPath(id), {
+      url,
+      timestamp: Date.now(),
+      browser: 'iPhone 6',
+      network: '3G',
+      localtion: '厦门',
     });
+    try {
+      this.queue.push(id);
+      this.flushQueue();
+    } catch (error) {
+      // ignore
+    }
+    return id;
+  }
+
+  async getResult(
+    id: string,
+  ): Promise<{
+    stage: number; // -1|未知;0|成功;1|测试;>2|排队中
+    info?: object;
+    browsertime?: object;
+    har?: object;
+  }> {
+    let info = null;
+    try {
+      info = await fse.readJson(getInfoPath(id));
+    } catch (error) {}
+    if (info) {
+      try {
+        const info = await fse.readJson(getInfoPath(id));
+        const browsertimePath = getBrowsertimePath(id);
+        const isEnd = await fse.pathExists(browsertimePath);
+        if (isEnd) {
+          const browsertime = await fse.readJson(browsertimePath);
+          const har = await fse.readJson(getHARPath(id));
+          return {
+            stage: 0,
+            info,
+            browsertime,
+            har,
+          };
+        } else {
+          const count = this.getPendingCountById(id);
+          if (count >= 0) {
+            return {
+              stage: count + 1,
+            };
+          }
+        }
+      } catch (error) {
+        // TODO
+      }
+    }
+    return {
+      stage: -1,
+    };
   }
 }
